affichage planning pas de doublons et mieux affichés, correction du pb du fichier caché dans output>trips

+++ initialize.py
@@ -24,4 +24,5 @@ for file in datafiles:
         TypedSet(Trip)), f)
 
 for folder in os.listdir(path):
-    shutil.rmtree(f"{path}/{folder}")
+    if folder[0] != "." :
+        shutil.rmtree(f"{path}/{folder}")
